requirements.txt
dpkt==1.9.8
ethernetip==1.1.2
festo-edcon==0.15.0
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
pluggy==1.6.0
pygments==2.19.2
pymodbus==3.11.3
pytest==8.4.2
pytest-mock==3.15.1
rich==14.2.0
setuptools==80.9.0
